Move MockLocalLayoutStorage
